pyproject.toml
[project]
name = "blog-to-podcast-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.2.10",
    "alembic>=1.18.4",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "elevenlabs>=1.0.0",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.135.3",
    "firecrawl-py>=4.6.0",
    "greenlet>=3.3.2",
    "groq>=1.1.2",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pillow>=12.2.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.24",
    "sqlalchemy>=2.0.49",
]
